Update ACC support to use Fortran runtime functions, supporting Cray compilers
