+++ src/atlas/array/gridtools/GridToolsArray.cc
@@ -25,7 +25,7 @@
 #include "atlas/runtime/Log.h"
 
 #if ATLAS_HAVE_ACC
-#include "atlas_acc_support/atlas_acc_map_data.h"
+#include "atlas_acc_support/atlas_acc.h"
 #endif
 
 //------------------------------------------------------------------------------
