+++ src/atlas/parallel/acc/acc.cc
@@ -13,7 +13,20 @@
 #include "atlas/library/defines.h"
 
 #if ATLAS_HAVE_ACC
-#include "atlas_acc_support/atlas_acc_map_data.h"
+#include "hic/hic.h"
+#include "atlas_acc_support/atlas_acc.h"
+static int hic_devices() {
+    static int devices_ = [](){
+        int n = 0;
+        auto err = hicGetDeviceCount(&n);
+        if (err != hicSuccess) {
+            n = 0;
+            static_cast<void>(hicGetLastError());
+        }
+        return n;
+    }();
+    return devices_;
+}
 #endif
 
 namespace atlas::acc {
@@ -21,6 +34,9 @@ namespace atlas::acc {
 int devices() {
 #if ATLAS_HAVE_ACC
     static int num_devices = [](){
+        if (hic_devices() == 0) {
+            return 0;
+        }
         auto devicetype = atlas_acc_get_device_type();
         int _num_devices = atlas_acc_get_num_devices();
         if (_num_devices == 1 && devicetype == atlas_acc_device_host) {
