+++ src/atlas_acc_support/atlas_acc.h
@@ -10,6 +10,7 @@
 
 #pragma once
 
+#include <stddef.h>
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -19,18 +20,11 @@ typedef enum {
     atlas_acc_device_not_host = 1
 } atlas_acc_device_t;
 
-void atlas_acc_map_data(void* cpu_ptr, void* gpu_ptr, unsigned long bytes);
+void atlas_acc_map_data(void* cpu_ptr, void* gpu_ptr, size_t bytes);
 void atlas_acc_unmap_data(void* cpu_ptr);
-int atlas_acc_is_present(void* cpu_ptr, unsigned long bytes);
+int atlas_acc_is_present(void* cpu_ptr, size_t bytes);
 void* atlas_acc_deviceptr(void* cpu_ptr);
 atlas_acc_device_t atlas_acc_get_device_type();
-
-int atlas_acc_devices();
-
-const char* atlas_acc_version_str();
-
-const char* atlas_acc_info_str();
-
 int atlas_acc_get_num_devices();
 
 #ifdef __cplusplus
