+++ src/sandbox/fortran_acc_fields/CMakeLists.txt
@@ -10,10 +10,8 @@ if( atlas_HAVE_ACC )
     ecbuild_add_executable(
         TARGET  atlas-acc-fields
         SOURCES atlas-acc-fields.F90
-        LIBS    atlas_f
+        LIBS    atlas_f OpenACC::OpenACC_Fortran
         LINKER_LANGUAGE Fortran
         NOINSTALL
     )
-    target_compile_options( atlas-acc-fields PUBLIC ${ACC_Fortran_FLAGS} )
-    target_link_libraries( atlas-acc-fields ${ACC_Fortran_FLAGS} )
 endif()
