+++ src/tests/acc/CMakeLists.txt
@@ -6,33 +6,25 @@
 # granted to it by virtue of its status as an intergovernmental organisation nor
 # does it submit to any jurisdiction.
 
-if( HAVE_CUDA AND HAVE_TESTS AND HAVE_FCTEST AND HAVE_ACC )
+if( HAVE_GPU AND HAVE_TESTS AND HAVE_FCTEST AND HAVE_ACC )
 
-  string (REPLACE ";" " " ACC_Fortran_FLAGS_STR "${ACC_Fortran_FLAGS}")
-
-
-  set_source_files_properties( fctest_unified_memory_with_openacc.F90 PROPERTIES COMPILE_FLAGS ${ACC_Fortran_FLAGS_STR} )
   add_fctest( 
     TARGET atlas_test_unified_memory_with_openacc
     SOURCES fctest_unified_memory_with_openacc.F90
             fctest_unified_memory_with_openacc_cxx.cc
-    LIBS atlas_f
+    LIBS atlas_f OpenACC::OpenACC_Fortran hic
     LINKER_LANGUAGE Fortran
     ENVIRONMENT ${ATLAS_TEST_ENVIRONMENT} ATLAS_RUN_NGPUS=1
   )
-  target_link_libraries( atlas_test_unified_memory_with_openacc ${ACC_Fortran_FLAGS} hic )
-
   set_tests_properties( atlas_test_unified_memory_with_openacc PROPERTIES LABELS "gpu;acc")
 
   add_fctest( 
     TARGET atlas_test_connectivity_openacc
     SOURCES fctest_connectivity_openacc.F90
-    LIBS atlas_f
+    LIBS atlas_f OpenACC::OpenACC_Fortran
     LINKER_LANGUAGE Fortran
     ENVIRONMENT ${ATLAS_TEST_ENVIRONMENT} ATLAS_RUN_NGPUS=1
   )
-  target_link_libraries( atlas_test_connectivity_openacc ${ACC_Fortran_FLAGS} )
-  set_target_properties( atlas_test_connectivity_openacc PROPERTIES COMPILE_FLAGS "${ACC_Fortran_FLAGS_STR}" )
   set_tests_properties ( atlas_test_connectivity_openacc PROPERTIES LABELS "gpu;acc")
 
 endif()
