+++ src/tests/array/test_array.cc
@@ -19,7 +19,6 @@
 #include "atlas/array/gridtools/GridToolsMakeView.h"
 #endif
 
-#include "hic/hic.h"
 #include "atlas/parallel/acc/acc.h"
 
 using namespace atlas::array;
@@ -556,25 +555,12 @@ CASE("test_wrap") {
     EXPECT(view(2) == 19);
 }
 
-static int hic_devices() {
-    static int devices_ = [](){
-        int n = 0;
-        auto err = hicGetDeviceCount(&n);
-        if (err != hicSuccess) {
-            n = 0;
-            static_cast<void>(hicGetLastError());
-        }
-        return n;
-    }();
-    return devices_;
-}
-
 CASE("test_acc_map") {
     Array* ds = Array::create<double>(2, 3, 4);
     EXPECT_NO_THROW(ds->allocateDevice());
     if( ds->deviceAllocated() ) {
         EXPECT_NO_THROW(ds->accMap());
-        EXPECT_EQ(ds->accMapped(), std::min(std::min(acc::devices(),hic_devices()),1));
+        EXPECT_EQ(ds->accMapped(), std::min(acc::devices(),1));
     }
     else {
         Log::warning() << "WARNING: Array could not be allocated on device, so acc_map could not be tested" << std::endl;
