+++ src/tests/field/CMakeLists.txt
@@ -27,13 +27,11 @@ ecbuild_add_test( TARGET atlas_test_field_foreach
 
 ecbuild_add_test( TARGET atlas_test_field_acc
   SOURCES  test_field_acc.cc
-  LIBS     atlas
+  LIBS     atlas OpenACC::OpenACC_CXX
   ENVIRONMENT ${ATLAS_TEST_ENVIRONMENT}
-  CONDITION atlas_HAVE_ACC
+  CONDITION atlas_HAVE_ACC AND OpenACC_CXX_Found
 )
 if( TEST atlas_test_field_acc )
-    target_compile_options( atlas_test_field_acc PRIVATE "${ACC_C_FLAGS}")
-    target_link_options( atlas_test_field_acc PRIVATE "${ACC_C_FLAGS}")
     set_tests_properties ( atlas_test_field_acc PROPERTIES LABELS "gpu;acc")
 endif()
 
@@ -73,14 +71,11 @@ if( HAVE_FCTEST )
     CONDITION       atlas_HAVE_ACC
     LINKER_LANGUAGE Fortran
     SOURCES         fctest_field_gpu.F90 external_acc_routine.F90
-    LIBS            atlas_f
+    LIBS            atlas_f OpenACC::OpenACC_Fortran
     ENVIRONMENT     ${ATLAS_TEST_ENVIRONMENT} ATLAS_RUN_NGPUS=1
   )
 
   if( TARGET atlas_fctest_field_device )
-    target_compile_options( atlas_fctest_field_device PUBLIC ${ACC_Fortran_FLAGS} )
-    target_link_libraries( atlas_fctest_field_device ${ACC_Fortran_FLAGS} )
-    target_link_options( atlas_fctest_field_device PRIVATE "${ACC_Fortran_FLAGS}")
     set_tests_properties ( atlas_fctest_field_device PROPERTIES LABELS "gpu;acc")
   endif()
 
@@ -88,14 +83,11 @@ if( HAVE_FCTEST )
     CONDITION       atlas_HAVE_ACC
     LINKER_LANGUAGE Fortran
     SOURCES         fctest_field_wrap_gpu.F90 external_acc_routine.F90
-    LIBS            atlas_f
+    LIBS            atlas_f OpenACC::OpenACC_Fortran
     ENVIRONMENT     ${ATLAS_TEST_ENVIRONMENT} ATLAS_RUN_NGPUS=1
   )
 
   if( TARGET atlas_fctest_field_wrap_device )
-    target_compile_options( atlas_fctest_field_wrap_device PUBLIC ${ACC_Fortran_FLAGS} )
-    target_link_libraries( atlas_fctest_field_wrap_device ${ACC_Fortran_FLAGS} )
-    target_link_options( atlas_fctest_field_wrap_device PRIVATE "${ACC_Fortran_FLAGS}")
     set_tests_properties ( atlas_fctest_field_wrap_device PROPERTIES LABELS "gpu;acc")
   endif()
